requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.6
python-dotenv==1.0.1
requests==2.32.4
